requirements.txt
python-dotenv==1.0.0
pydantic==2.5.2
pandas==2.0.3
numpy==1.26.4
MetaTrader5==5.0.4424
python-telegram-bot==20.8